PP-10496 Rename smokeTestHelpers file
- Renamed smokeTestHelpers file to follow our style guide for file names
