+++ helpers/smoke-test-helpers.js
