+++ make-card-payment-sandbox-without-3ds/index.js
@@ -1,7 +1,7 @@
 const { ENVIRONMENT, WEBHOOKS_ENABLED } = process.env
 
 const log = require('SyntheticsLogger')
-const smokeTestHelpers = require('../helpers/smokeTestHelpers')
+const smokeTestHelpers = require('../helpers/smoke-test-helpers')
 
 const sandboxCard = {
   cardholderName: 'Test User',
